+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Added
 * Added `check` kwarg to `ub.cmd`, which when True will raise a
   `CalledProcessError` if the exit-code is non-zero.
+* Added support for pypy.
 
 ### Changed
 * Moved `timerit` to its own module.